feat(movie-card-tabs): интеграция компонента табов в страницу с фильмами
